+++ R/identifier_exists.R
@@ -10,14 +10,13 @@
 #' @importFrom methods is
 #' 
 #' @return Logical, \code{TRUE} or \code{FALSE}
-#' @examples
-#' 
+#' @examples \donttest{
 #' # identifier exists
 #' identifier_exists(sb_id = "4f4e4b24e4b07f02db6aea14")
 #' 
 #' # identifier does not exist
 #' identifier_exists(sb_id = "aaaaaaakkkkkkkbbbbbb")
-#' 
+#' }
 identifier_exists <- function(sb_id, ..., session = current_session()) {
 	#sb_id = as.sbitem(sb_id)
 	if(is(sb_id, 'sbitem')){