+++ R/item_get_fields.R
@@ -7,8 +7,7 @@
 #' @return List serialization of chosen metadata for an SB item
 #' @import httr
 #' @export
-#' @examples 
-#' \donttest{
+#' @examples \donttest{
 #' # Get certain fields from an item
 #' item_get_fields("4f4e4b24e4b07f02db6aea14", c('title', 'citation', 'contacts'))
 #'