+++ R/item_rm_files.R
@@ -1,8 +1,15 @@
 #' @title Remove files associated with an item
 #' 
 #' @template manipulate_item
-#' @param files A character vector of file names to remove. If not supplied, defaults to removing all attached files.
+#' @param files A character vector of file names to remove. 
+#' If not supplied, defaults to removing all attached files.
+#' 
 #' @description Removes existing files associated with an item.
+#' 
+#' NOTE: This function will not alter facets which can also
+#' contain facets. To manipulate facets, the facet element of
+#' a sciencebase item must be altered and updated with \code{\link{item_update}}.
+#' 
 #' @return An updated object of class \code{sbitem}
 #' @description 
 #' This function is the key way to remove files attached to SB items. 
@@ -38,7 +45,11 @@ item_rm_files <- function(sb_id, files,..., session=current_session()){
 		files_to_keep = item$files[!fnames %in% basename(files)]
 		#files_to_keep = lapply(files_to_keep, function(x){x[c('name', 'title', 'contentType')]})
 	}
+
+	if(length(files_to_keep) == 0 && is.list(files_to_keep)) {
+		return(item)
+	}
 	
 	as.sbitem(item_update(item$id, info = list(files = files_to_keep), ..., session = session))
-	
+
 }