+++ R/query_sb_text.R
@@ -1,6 +1,4 @@
 #' @title Query SB for items containing specific text
-#' 
-#' 
 #' @param text Text string for search
 #' @inheritParams query_sb
 #' 
@@ -12,13 +10,13 @@
 #' Queries for ScienceBase items that have matching text in the title or 
 #' description
 #' 
-#' @examples 
+#' @examples \donttest{
 #' #query for a person's name
 #' query_sb_text('Luna Leopold')
 #' 
 #' #query for one of the old river gaging stations
 #' query_sb_text('Lees Ferry')
-#' 
+#' }
 #' 
 #' @export
 query_sb_text = function(text, ..., limit=20, session=current_session()){